Change indentation errors (again)
We now emit one error for each group of contiguous lines which have the same indentation mistake (i.e. the same common prefix with the expected indentation). That error mentions the number of lines and replaces only whatever indentation is common among those lines.

In cases where the entire literal is indented consistently with itself, but not with the closing delimiter, this should give you a single error covering the entire literal, with a single fix-it that consistently reindents the entire literal. In more complex cases, it still keeps the number of error messages manageable.