[Foundation] Improve String bridge performance by implementing SPIs to access character buffers directly
When Swift strings are bridged back into objective-c two distinct methods can be called to get fast-access to the underlying buffers of the strings. These come in two flavors - either eight-bit contents or unicode contents. When the backing of the swift string is represented as a buffer of unicode characters (an element width of 2) the fastest and smallest encoding is unicode (NSUnicodeStringEncoding = 10) and the buffer can be portrayed to obejctive-c directly as a cast of the startUTF16 to a constant unichar pointer. Likewise when the backing is ASCII the buffer can be represented as an eight bit encoding (NSASCIIStringEncoding = 1) and the buffer can be directly accessed as a constant char pointer.

This in normal cases can improve the bridging performance in the upwards of 90% faster access in bridged cases when passing a swift string into objective-c APIs.