Short circuit subspace check with empty subjects
When determining whether one constructor pattern is a subspace of
another, we only checked if the right-hand subspace was larger than the
left.  If the left is larger than the right then we can short-circuit
the comparison once more because it cannot be a subspace.

This also takes care of refutable casts in patterns tripping the
duplicate case warning.