Diagnose Substring-to-String conversions resulting from subscripting.
When we have a Substring-to-String conversion as a result of
subscripting in a context expecting String, we emit the rather generic
and unactionable availability diagnostic.

Instead, to aid in migration, emit a highly specialized diagnostic and
fixit when we're compiling for -swift-version 4.

(cherry picked from commit 70c0e855819280a2ce8d8ba8aed7634a36454b57)