Revert "Re-instate "Improve disjunction selection""
This reverts commit
ecfa406fc580df48260c6ad606bfd11c10f6ad4c, which was reinstating
3c5b393e0c07c222950e626e5c4553358e22365d.

It does not revert one test change from that commit, because
inexplicably one of the tests is still failing, probably due to some
other changes that have happened since. I'm leaving a ticket open to
investigate.

This is being reverted not because of bugs in this particular commit,
but because it seems to be exposing other bugs in the type checker that
are resulting in source compatibility problems. We need to shake these
other bugs out of the type checker before bringing this change back.

Fixes rdar://problem/32204276 and a few dups.

(cherry picked from commit bc7dff5c9f1a861682eadd87e6631b1c78540cb4)