Revert "Re-instate "Improve disjunction selection""
This reverts commit
ecfa406fc580df48260c6ad606bfd11c10f6ad4c, which was reinstating
3c5b393e0c07c222950e626e5c4553358e22365d.

It does not revert one test change from that commit, because
inexplicably one of the tests is still failing, probably due to some
other changes that have happened since. I'm leaving a ticket open to
investigate.

I've added one of the new cases that is failing as a result of this
change.

This is being reverted not because of bugs in this particular commit,
but because it seems to be exposing other bugs in the type checker that
are resulting in source compatibility problems. We need to shake these
other bugs out of the type checker before bringing this change back.