[GSB] Only finalize a canonical GSB when performing expensive assertions.
A `GenericSignatureBuilder` only needs to be finalized when we need to
compute a generic signature from it or otherwise require diagnostics.
Canonical generic signature builders don’t need either of those, so unless we
are performing the expensive idempotency checking, only do the minimal work
to “process delayed requirements”.

This gives a 3x speedup in type-checking time for the “trivial” example

    let x = [1, 2]

because we do a lot less work on deserialization of generic environments.

Part of rdar://problem/32116933.
(cherry picked from commit 4f8f86ab2f4521bf70f42822d7ea1232429d5474)