Remove test added with revert a1ed4e61748dbdc7962979c04922becd68cbb3ab.
I added a test case from one of the failures that were exposed by the
change that I was reverting in
https://github.com/apple/swift/commit/a1ed4e61748dbdc7962979c04922becd68cbb3ab,
but I was unable to make this test not dependent on CGFloat. We don't
have that available everywhere, and the mock SDK doesn't support
everything we need here, so I'm just going to remove the test since
adding it isn't absolutely necessary at this time.