[ClangImporter] Import builtins returning size_t as returning Int. (#9753)
...which is consistent with other library functions return size_t. But
don't break compatibility with Swift 3.

("Builtins" are functions that Clang has, well, built-in knowledge
about, and whose declarations can therefore be produced directly in
the compiler. This bypassed some of our checking for special typedefs
like size_t. We discovered this when 'wcslen' got added as a new
built-in in the Swift 4 timeframe.)

rdar://problem/30545505