Sema: Better handling of type members in getTypeOfMemberReference()
We don't actually want to open the type; instead, calculate the
substituted member type using substMemberTypeWithBase(), which
will return an unbound generic type if the member is a generic
nominal or type alias, and then it open it with openUnboundGenericType().

This is NFC for now, but I need generic typealiases to come in as
UnboundGenericTypes here in a subsequent patch.