Sema: Don't allow unbound generic types to leak into types of VarDecls
Stored property types can contain unbound generic types
as long as there is an initializer expression.

However, if type checking the initializer expression
failed, we didn't set the type of the VarDecl to an
error type.

This fixes some regressions with an upcoming patch that
tightens invariants around unbound generic types.