[ConstraintSystem] Invalidate solution in presence of error types
In `ConstraintSystem::finalize` if simplification of one of the
type variables present in the system results in introduction of
error type to the possible solution, let's not return that solution,
because that might cause problems down the road in verifier.

One of the reasons why `simplifyType` could produce an error type
is related to substitution of associated types using newly
discovered base type because its protocol conformance as not
properly verified until later on.

Resolves: SR-4575.