[stdlib] Heterogeneous comparisons for concrete integer types
With the introduction of heterogeneous overloads for equality and
comparison operators, it became possible to compare UInt to Int8 and
vice versa, but it also led to the problems like the one reported in:

https://bugs.swift.org/browse/SR-4984

```
(swift) UInt() != 0xFFFF_FFFF_FFFF_FFFF
<REPL Input>:1:11: error: integer literal '18446744073709551615' overflows when stored into 'Int'
UInt() != 0xFFFF_FFFF_FFFF_FFFF
          ^
```

In this case even in the presence of a homogeneous overload for `!=`,
the typechecker first tries `Int` as a type for a literal and
successfully finds a solution for the whole expression via a
heterogeneous overload.

At the same time `UInt() == 0XFFFF_FFFF_FFFF_FFFF` works, i.e. selects
the correct overload. Which means that homogeneous versions of derived
equality and comparison operators should be moved to concrete types, in
order to be preferred by the solver. And this is exactly what this
change is doing.