Eliminate most uses of ArchetypeResolutionKind::AlwaysPartial.
When resolving archetypes using the "always partial" resolution kind,
we allow the system to form potential archetypes which might be
invalid. Start limiting the use of this "always partial" resolution
kind, so that it can eventually be removed entirely to simplify the
invariants of the GenericSignatureBuilder.

Each of the callers should either be working with complete,
well-formed archetypes or they should not force any resolution this
early.