In markdown break test names with zero-width space
In the `--markdown` format, which is used to display performance test comparisons on GitHub, break the long test names on the camel case word boundary using zero-width space.

This prevents the results table from overflowing badly, especially in mobile view. But the **downside** is that you can no longer simply find the tests in the project by the names copied from comparison results, because they now contain invisible characters.

Use of zero-with space required the move of some strings to unicode, otherwise the column justification and formatting broke on multibyte characters in UTF-8.

Also updated class declarations and private methods to better follow Python conventions. And used list comprehension where it is clearer than map.

Added tests for parts of compare_perf_tests.py and their `lit` invocation to the `validation-test`.